requirements.txt
numpy
pyopencl
colorama
pillow
